[#96] Refactor: separate DeletAccountSuccessModal into file
