+++ src/pages/DeleteAccount/DeleteAccountSuccessModal.tsx
@@ -1,0 +1,42 @@
+import { useSetAtom } from 'jotai';
+import { useNavigate } from 'react-router-dom';
+
+import { Modal } from '@/components/Modal';
+import { ModalButton } from '@/components/Modal/ModalButton';
+import { MainText, SubText, ModalText } from '@/components/Modal/ModalText';
+import { ROUTE_PATH } from '@/constants';
+import { loginStateAtom } from '@/stores/atoms/loginStateAtom';
+import { userDataAtom } from '@/stores/atoms/userDataAtom';
+
+import { TEXT_CONTENTS } from './constants';
+
+import type { Dispatch, SetStateAction } from 'react';
+
+type DeleteAccountSuccessModalProps = {
+  setShowModal: Dispatch<SetStateAction<boolean>>;
+};
+
+export const DeleteAccountSuccessModal = ({ setShowModal }: DeleteAccountSuccessModalProps) => {
+  const setLoginState = useSetAtom(loginStateAtom);
+  const setUserData = useSetAtom(userDataAtom);
+  const navigate = useNavigate();
+
+  const handleOkButtonClick = () => {
+    setShowModal(false);
+    setLoginState(false);
+    setUserData(null);
+    navigate(ROUTE_PATH.roadmap.index);
+  };
+
+  return (
+    <Modal>
+      <ModalText>
+        <MainText textStyle="border-b-2 border-[#686868] py-5">
+          {TEXT_CONTENTS.modal.title}
+        </MainText>
+        <SubText textStyle="pt-6 pb-7">{TEXT_CONTENTS.modal.greeting}</SubText>
+      </ModalText>
+      <ModalButton text="확인" handler={handleOkButtonClick} isPrimary />
+    </Modal>
+  );
+};
